tried multiplying bitmap cost
we used to add some additional cost to a bitmap, but if we have 2 small bitmaps vs 1 big bitmap and we prefer the smaller two, this could potentially make the two have a higher cost than the big one. However, if we multiply two sides of the inequality, this will no longer be a concern.